Removed scala dependencies from the compiler project
This has to be done because Scala has a real problem around binary incompatibility
of different Scala versions.

Up until now Mustache.java has been lucky enough that the classes it uses were
only deprecated in Scala 2.10 (e.g. [ScalaObject] and [JavaConversions.asJavaMap]).

In Scala 2.11 these interfaces have been removed. In order to support Scala in
the future I propose removing all the Scala code from the compiler project and
creating a separate project that can be published in multiple forms, one for
each of the scala versions that are commonly used. I'd propose Scala 2.9, 2.10
and 2.11.

  [ScalaObject]: http://www.scala-lang.org/api/2.10.4/index.html#scala.ScalaObject
  [JavaConversions.asJavaMap]: http://www.scala-lang.org/api/2.11.1/index.html#scala.collection.JavaConversions$